+++ src/main/java/com/bside/sidefriends/checklist/service/ChecklistServiceImpl.java
@@ -10,11 +10,11 @@ import com.bside.sidefriends.checklist.repository.ChecklistMetaRepository;
 import com.bside.sidefriends.checklist.repository.ChecklistRepository;
 import com.bside.sidefriends.checklist.service.dto.*;
 import com.bside.sidefriends.pet.domain.Pet;
-import com.bside.sidefriends.pet.error.exception.PetDeactivatedException;
-import com.bside.sidefriends.pet.error.exception.PetNotFoundException;
+import com.bside.sidefriends.pet.exception.PetDeactivatedException;
+import com.bside.sidefriends.pet.exception.PetNotFoundException;
 import com.bside.sidefriends.pet.repository.PetRepository;
 import com.bside.sidefriends.users.domain.User;
-import com.bside.sidefriends.users.error.exception.UserMainPetNotFoundException;
+import com.bside.sidefriends.users.exception.UserMainPetNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
