--- src/main/java/com/bside/sidefriends/family/error/exception/FamilyDeleteFailException.java
@@ -1,14 +1,0 @@
-package com.bside.sidefriends.family.error.exception;
-
-import com.bside.sidefriends.common.exception.BusinessException;
-import com.bside.sidefriends.common.response.ResponseCode;
-
-public class FamilyDeleteFailException extends BusinessException {
-    public FamilyDeleteFailException() {
-        super(ResponseCode.F_DELETE_NOT_ALLOWED);
-    }
-
-    public FamilyDeleteFailException(String message) {
-        super(message, ResponseCode.F_DELETE_NOT_ALLOWED);
-    }
-}
