--- src/main/java/com/bside/sidefriends/family/error/exception/FamilyLimitExceededException.java
@@ -1,11 +1,0 @@
-package com.bside.sidefriends.family.error.exception;
-
-import com.bside.sidefriends.common.exception.BusinessException;
-import com.bside.sidefriends.common.response.ResponseCode;
-
-public class FamilyLimitExceededException extends BusinessException {
-
-    public FamilyLimitExceededException() {
-        super(ResponseCode.F_MAX_NUMBER_EXCEEDED);
-    }
-}
