--- src/main/java/com/bside/sidefriends/family/error/exception/FamilyManagerNotFoundException.java
@@ -1,11 +1,0 @@
-package com.bside.sidefriends.family.error.exception;
-
-import com.bside.sidefriends.common.exception.BusinessException;
-import com.bside.sidefriends.common.response.ResponseCode;
-
-public class FamilyManagerNotFoundException extends BusinessException {
-
-    public FamilyManagerNotFoundException(BusinessException businessException) {
-        super(ResponseCode.F_MANAGER_NOT_FOUND, businessException);
-    }
-}
