--- src/main/java/com/bside/sidefriends/family/error/exception/FamilyMemberDeleteFailException.java
@@ -1,16 +1,0 @@
-package com.bside.sidefriends.family.error.exception;
-
-import com.bside.sidefriends.common.exception.BusinessException;
-import com.bside.sidefriends.common.response.ResponseCode;
-
-public class FamilyMemberDeleteFailException extends BusinessException {
-
-    public FamilyMemberDeleteFailException(String message) {
-        super(message, ResponseCode.F_DELETE_MEMBER_NOT_ALLOWED);
-    }
-
-    public FamilyMemberDeleteFailException(BusinessException businessException) {
-        super(ResponseCode.F_DELETE_MEMBER_NOT_ALLOWED, businessException);
-    }
-
-}
