--- src/main/java/com/bside/sidefriends/family/error/exception/FamilyMemberNotFoundException.java
@@ -1,12 +1,0 @@
-package com.bside.sidefriends.family.error.exception;
-
-import com.bside.sidefriends.common.exception.BusinessException;
-import com.bside.sidefriends.common.response.ResponseCode;
-
-public class FamilyMemberNotFoundException extends BusinessException {
-
-    public FamilyMemberNotFoundException(BusinessException businessException) {
-        super(ResponseCode.F_MEMBER_NOT_FOUND, businessException);
-    }
-
-}
