--- src/main/java/com/bside/sidefriends/pet/error/exception/PetDeactivatedException.java
@@ -1,8 +1,0 @@
-package com.bside.sidefriends.pet.error.exception;
-
-import com.bside.sidefriends.common.exception.BusinessException;
-import com.bside.sidefriends.common.response.ResponseCode;
-
-public class PetDeactivatedException extends BusinessException {
-    public PetDeactivatedException() { super(ResponseCode.P_ENTITY_DEACTIVATED);}
-}
