--- src/main/java/com/bside/sidefriends/pet/error/exception/PetShareFailException.java
@@ -1,11 +1,0 @@
-package com.bside.sidefriends.pet.error.exception;
-
-import com.bside.sidefriends.common.exception.BusinessException;
-import com.bside.sidefriends.common.response.ResponseCode;
-
-public class PetShareFailException extends BusinessException {
-
-    public PetShareFailException(String message) {
-        super(message, ResponseCode.P_SHARE_FAIL);
-    }
-}
